Move domain suggestions store to stores/
