+++ client/landing/gutenboarding/stores/domain-suggestions/constants.ts
