+++ client/landing/gutenboarding/stores/domain-suggestions/types.ts
